requirements.txt
asgiref==3.5.1
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.8
pytz==2022.1
shortuuid==1.0.8
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
